get bikes data in component
