+++ src/app/containers/bike-list/bike-list-container.component.html
@@ -1,1 +1,2 @@
-<app-bike-list></app-bike-list>
+<app-bike-list [bikesList]="bikesList$ | async"
+(searchBikes)="searchBikes($event)"></app-bike-list>
