+++ src/app/store/app-state.model.ts
@@ -1,0 +1,5 @@
+import {Bike} from "../model/bike.model";
+
+export interface AppState {
+    readonly bikes: Array<Bike>;
+}
