--- src/app/store/state.model.ts
@@ -1,6 +1,0 @@
-import { Bike } from "../model/bike.model";
-
-export interface AppState {
-    readonly bikes: Array<Bike>;
-    readonly users: Bike;
-}
